Parse s6-linux-init arguments before execing
Execing into s6-linux-init-telinit results in the USAGE message of
s6-linux-init to be shown if it is actually running as pid1 already,
which makes it significantly less usefull. By parsing the command line
first, we allow the user to use arguments like -h to produce
the acutally expected usage message of s6-linux-init instead of
s6-linux-init-telinit.

Signed-off-by: Christian Kampka <christian@kampka.net>